Refactor: Replace `Int` with `BigInteger` for amount and balance fields
- Updated spin use cases, events, and models to use `BigInteger` for precise amount handling.
- Adjusted database mappings and repository logic for `BigInteger` support.
- Refactored `FakePlayerAdapter`, `FakeWalletAdapter`, and domain errors to align with `BigInteger` changes.
- Updated event serialization with `@Contextual` annotation for `BigInteger` fields.

+++ src/main/kotlin/application/port/outbound/PlayerAdapter.kt
@@ -1,5 +1,7 @@
 package application.port.outbound
 
+import java.math.BigInteger
+
 /**
  * Port interface for player operations.
  * Implementations connect to external player service.
@@ -9,5 +11,5 @@ interface PlayerAdapter {
      * Get player's current bet limit.
      * Returns null if no limit is set.
      */
-    suspend fun findCurrentBetLimit(playerId: String): Result<Int?>
+    suspend fun findCurrentBetLimit(playerId: String): Result<BigInteger?>
 }
