+++ src/main/kotlin/infrastructure/aggregator/onegamehub/handler/OneGameHubRoute.kt
@@ -9,7 +9,7 @@ import org.koin.ktor.ext.inject
 import shared.value.SessionToken
 import kotlin.getValue
 
-private val Parameters.amount get() = this["amount"]!!.toInt()
+private val Parameters.amount get() = this["amount"]!!.toBigInteger()
 private val Parameters.gameSymbol get() = this["game_id"]!!
 private val Parameters.transactionId get() = this["transaction_id"]!!
 private val Parameters.roundId get() = this["round_id"]!!
