+++ src/main/kotlin/infrastructure/persistence/exposed/repository/ExposedGameWonRepository.kt
@@ -4,6 +4,7 @@ import domain.game.repository.GameWonRepository
 import infrastructure.persistence.exposed.table.GameWonTable
 import org.jetbrains.exposed.sql.insert
 import org.jetbrains.exposed.sql.transactions.experimental.newSuspendedTransaction
+import java.math.BigInteger
 import java.util.UUID
 
 /**
@@ -11,12 +12,12 @@ import java.util.UUID
  */
 class ExposedGameWonRepository : GameWonRepository {
 
-    override suspend fun save(gameId: UUID, playerId: String, amount: Int, currency: String): Boolean =
+    override suspend fun save(gameId: UUID, playerId: String, amount: BigInteger, currency: String): Boolean =
         newSuspendedTransaction {
             GameWonTable.insert {
                 it[GameWonTable.gameId] = gameId
                 it[GameWonTable.playerId] = playerId
-                it[GameWonTable.amount] = amount
+                it[GameWonTable.amount] = amount.toLong()
                 it[GameWonTable.currency] = currency
             }
             true
